+++ src/components/public/gnb/components/ProfileComponent.tsx
@@ -5,8 +5,7 @@ import { useRouter } from "next/navigation"
 
 import { useRef, useState } from "react"
 
-import LoginModal from "@/components/pages/auth/LoginModal"
-import SignUpModal from "@/components/pages/auth/SignUpModal/SignUpModal"
+import AuthModal from "@/components/pages/auth/AuthModal"
 import Profile from "@/components/public/img/Profile"
 import ROUTE from "@/constants/route"
 import useOutsideClick from "@/util/useOutsideClick"
@@ -59,9 +58,7 @@ const SignIn = () => {
         />
         <p className="relative z-10">로그인</p>
       </button>
-
-      <SignUpModal />
-      <LoginModal isLogin={isLogin} setIsLogin={setIsLogin} />
+      <AuthModal isLogin={isLogin} setIsLogin={setIsLogin} />
     </>
   )
 }
