+++ src/hooks/Auth/usePostSignup.tsx
@@ -3,11 +3,14 @@ import { Dispatch, SetStateAction } from "react"
 import signup from "@/actions/Auth-actions/signup"
 import { useMutation } from "@tanstack/react-query"
 
-const usePostSignup = (setError: Dispatch<SetStateAction<string>>) => {
+const usePostSignup = (
+  setError: Dispatch<SetStateAction<string>>,
+  setIsStep: Dispatch<SetStateAction<number>>,
+) => {
   return useMutation({
     mutationFn: signup,
     onSuccess: () => {
-      alert("회원가입 되었습니다.")
+      setIsStep(2)
     },
     onError: (error) => {
       setError(error.message)
