+++ src/types/auth/auth.d.ts
@@ -53,6 +53,13 @@ export interface IEmailExists extends IMessage {
 
 export type TSignupResponse = IEmailExists | IValidationError | IMessage
 
+// 리팩토링 후
+
+export interface PasswordVisibility {
+  showPwd: boolean
+  verifyShowPwd: boolean
+}
+
 export declare module "next-auth" {
   interface User {
     accessToken: string
